requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
nltk==3.8.1
sacrebleu==2.3.1
httpx==0.25.0
python-dotenv==1.0.0 